Now sometimes I get a black screen directly
Not even like before that the shape managed to show, I have went step by step, making sure it gets as close as needed, but still can't get to it really.

+++ Temp_Engine/FlyLib/flylib_src/fly_timer.cpp
@@ -12,9 +12,9 @@ typedef std::chrono::high_resolution_clock::duration high_res_clock;
 FLY_Timer::FLY_Timer() : is_stopped(false), is_active(true)
 {
     high_res_clock now = std::chrono::high_resolution_clock::now().time_since_epoch();
-    stop_at_ms = start_at_ms = (uint16_t)std::chrono::duration_cast<std::chrono::milliseconds>(now).count();
-    stop_at_us = start_at_us = (uint32_t)std::chrono::duration_cast<std::chrono::microseconds>(now).count();
-    stop_at_ns = start_at_ns = (uint32_t)std::chrono::duration_cast<std::chrono::nanoseconds>(now).count();
+    stop_at_ms = start_at_ms = (uint32)std::chrono::duration_cast<std::chrono::milliseconds>(now).count();
+    stop_at_us = start_at_us = (uint32)std::chrono::duration_cast<std::chrono::microseconds>(now).count();
+    stop_at_ns = start_at_ns = (uint32)std::chrono::duration_cast<std::chrono::nanoseconds>(now).count();
 }
 
 void FLY_Timer::Start()
@@ -22,17 +22,17 @@ void FLY_Timer::Start()
     is_active = true;
     is_stopped = false;
     high_res_clock now = std::chrono::high_resolution_clock::now().time_since_epoch();
-    stop_at_ms = start_at_ms = (uint16_t)std::chrono::duration_cast<std::chrono::milliseconds>(now).count();
-    stop_at_us = start_at_us = (uint32_t)std::chrono::duration_cast<std::chrono::microseconds>(now).count();
-    stop_at_ns = start_at_ns = (uint32_t)std::chrono::duration_cast<std::chrono::nanoseconds>(now).count();
+    stop_at_ms = start_at_ms = (uint32)std::chrono::duration_cast<std::chrono::milliseconds>(now).count();
+    stop_at_us = start_at_us = (uint32)std::chrono::duration_cast<std::chrono::microseconds>(now).count();
+    stop_at_ns = start_at_ns = (uint32)std::chrono::duration_cast<std::chrono::nanoseconds>(now).count();
 }
 
 void FLY_Timer::Stop()
 {
     high_res_clock now = std::chrono::high_resolution_clock::now().time_since_epoch();
-    stop_at_ms = (uint16_t)std::chrono::duration_cast<std::chrono::milliseconds>(now).count();
-    stop_at_us = (uint32_t)std::chrono::duration_cast<std::chrono::microseconds>(now).count();
-    stop_at_ns = (uint32_t)std::chrono::duration_cast<std::chrono::nanoseconds>(now).count();
+    stop_at_ms = (uint32)std::chrono::duration_cast<std::chrono::milliseconds>(now).count();
+    stop_at_us = (uint32)std::chrono::duration_cast<std::chrono::microseconds>(now).count();
+    stop_at_ns = (uint32)std::chrono::duration_cast<std::chrono::nanoseconds>(now).count();
     is_stopped = true;
 }
 
@@ -54,11 +54,11 @@ bool FLY_Timer::IsActive()
     return is_active;
 }
 
-uint16 FLY_Timer::ReadMilliSec()
+uint32 FLY_Timer::ReadMilliSec()
 {
     if (is_stopped || !is_active)
         return stop_at_ms - start_at_ms;
-    uint16 now = (uint16)std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::high_resolution_clock::now().time_since_epoch()).count();
+    uint32 now = (uint32)std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::high_resolution_clock::now().time_since_epoch()).count();
     return now - start_at_ms;
 }
